refactor: extract virtual-fs into utils
