+++ packages/create-cli/src/lib/setup/ci.unit.test.ts
@@ -1,9 +1,9 @@
 import { select } from '@inquirer/prompts';
 import { vol } from 'memfs';
 import { MEMFS_VOLUME } from '@code-pushup/test-utils';
+import { createTree } from '@code-pushup/utils';
 import { promptCiProvider, resolveCi } from './ci.js';
 import type { ConfigContext } from './types.js';
-import { createTree } from './virtual-fs.js';
 
 vi.mock('@inquirer/prompts', () => ({
   select: vi.fn(),
