+++ packages/utils/src/lib/wizard/index.ts
@@ -1,0 +1,2 @@
+export type { FileChange, FileSystemAdapter, Tree } from './types.js';
+export { createTree } from './virtual-fs.js';
